splitting out files to shared project
