+++ Core/Features/Markers/MarkerDto.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace LaHistoricalMarkers.Data
+namespace LaHistoricalMarkers.Core.Features.Markers
 {
     public class MarkerDto
     {
