+++ Core/Features/Markers/MarkerSubmissionDto.cs
@@ -1,11 +1,9 @@
 using System;
 
-namespace LaHistoricalMarkers.Data
+namespace LaHistoricalMarkers.Core.Features.Markers
 {
-    public class PendingSubmissionDto
+    public class MarkerSubmissionDto
     {
-        public int Id { get; set; }
-
         public string Name { get; set; }
 
         public string Description { get; set; }
@@ -14,6 +12,6 @@ namespace LaHistoricalMarkers.Data
 
         public decimal Longitude { get; set; }
 
-        public string ImageFileName { get; set; }
+        public string Base64Image { get; set; }
     }
 }
